consolidate ipv6 tests and cosmetic changes
